question restructure according to design
