+++ question/classes/local/bank/checkbox_column.php
@@ -15,61 +15,64 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * A column type for the name of the question name.
+ * A column with a checkbox for each question with name q{questionid}.
  *
  * @package   core_question
  * @copyright 2009 Tim Hunt
+ * @author    2021 Safat Shahin <safatshahin@catalyst-au.net>
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-
-namespace core_question\bank;
+namespace core_question\local\bank;
 defined('MOODLE_INTERNAL') || die();
+
+use core\output\checkbox_toggleall;
 
 
 /**
- * A column type for the name of the question name.
+ * A column with a checkbox for each question with name q{questionid}.
  *
  * @copyright 2009 Tim Hunt
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class question_name_column extends column_base {
-    protected $checkboxespresent = null;
+class checkbox_column extends column_base {
 
     public function get_name() {
-        return 'questionname';
+        return 'checkbox';
     }
 
     protected function get_title() {
-        return get_string('question');
+        global $OUTPUT;
+
+        $mastercheckbox = new checkbox_toggleall('qbank', true, [
+            'id' => 'qbheadercheckbox',
+            'name' => 'qbheadercheckbox',
+            'value' => '1',
+            'label' => get_string('selectall'),
+            'labelclasses' => 'accesshide',
+        ]);
+
+        return $OUTPUT->render($mastercheckbox);
     }
 
-    protected function label_for($question) {
-        if (is_null($this->checkboxespresent)) {
-            $this->checkboxespresent = $this->qbank->has_column('core_question\bank\checkbox_column');
-        }
-        if ($this->checkboxespresent) {
-            return 'checkq' . $question->id;
-        } else {
-            return '';
-        }
+    protected function get_title_tip() {
+        return get_string('selectquestionsforbulk', 'question');
     }
 
     protected function display_content($question, $rowclasses) {
-        $labelfor = $this->label_for($question);
-        if ($labelfor) {
-            echo '<label for="' . $labelfor . '">';
-        }
-        echo format_string($question->name);
-        if ($labelfor) {
-            echo '</label>';
-        }
+        global $OUTPUT;
+
+        $checkbox = new checkbox_toggleall('qbank', false, [
+            'id' => "checkq{$question->id}",
+            'name' => "q{$question->id}",
+            'value' => '1',
+            'label' => get_string('select'),
+            'labelclasses' => 'accesshide',
+        ]);
+
+        echo $OUTPUT->render($checkbox);
     }
 
     public function get_required_fields() {
-        return array('q.id', 'q.name');
-    }
-
-    public function is_sortable() {
-        return 'q.name';
+        return array('q.id');
     }
 }
