+++ question/classes/local/bank/menuable_action.php
@@ -19,10 +19,11 @@
  *
  * @package   core_question
  * @copyright 2019 Tim Hunt
+ * @author    2021 Safat Shahin <safatshahin@catalyst-au.net>
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace core_question\bank;
+namespace core_question\local\bank;
 defined('MOODLE_INTERNAL') || die();
