+++ question/classes/local/bank/modifier_name_column.php
@@ -15,60 +15,61 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * A column type for the name of the question creator.
+ * A column type for the name of the question last modifier.
  *
  * @package   core_question
  * @copyright 2009 Tim Hunt
+ * @author    2021 Safat Shahin <safatshahin@catalyst-au.net>
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace core_question\bank;
+namespace core_question\local\bank;
 defined('MOODLE_INTERNAL') || die();
 
 
 /**
- * A column type for the name of the question creator.
+ * A column type for the name of the question last modifier.
  *
  * @copyright 2009 Tim Hunt
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class creator_name_column extends column_base {
+class modifier_name_column extends column_base {
     public function get_name() {
-        return 'creatorname';
+        return 'modifiername';
     }
 
     protected function get_title() {
-        return get_string('createdby', 'question');
+        return get_string('lastmodifiedby', 'question');
     }
 
     protected function display_content($question, $rowclasses) {
-        if (!empty($question->creatorfirstname) && !empty($question->creatorlastname)) {
+        if (!empty($question->modifierfirstname) && !empty($question->modifierlastname)) {
             $u = new \stdClass();
-            $u = username_load_fields_from_object($u, $question, 'creator');
-            $date = userdate($question->timecreated, get_string('strftimedatetime', 'langconfig'));
+            $u = username_load_fields_from_object($u, $question, 'modifier');
+            $date = userdate($question->timemodified, get_string('strftimedatetime', 'langconfig'));
             echo fullname($u) . '<br>' . \html_writer::tag('span', $date, array('class' => 'date'));
         }
     }
 
     public function get_extra_joins() {
-        return array('uc' => 'LEFT JOIN {user} uc ON uc.id = q.createdby');
+        return array('um' => 'LEFT JOIN {user} um ON um.id = q.modifiedby');
     }
 
     public function get_required_fields() {
         $allnames = \core_user\fields::get_name_fields();
         $requiredfields = array();
         foreach ($allnames as $allname) {
-            $requiredfields[] = 'uc.' . $allname . ' AS creator' . $allname;
+            $requiredfields[] = 'um.' . $allname . ' AS modifier' . $allname;
         }
-        $requiredfields[] = 'q.timecreated';
+        $requiredfields[] = 'q.timemodified';
         return $requiredfields;
     }
 
     public function is_sortable() {
         return array(
-            'firstname' => array('field' => 'uc.firstname', 'title' => get_string('firstname')),
-            'lastname' => array('field' => 'uc.lastname', 'title' => get_string('lastname')),
-            'timecreated' => array('field' => 'q.timecreated', 'title' => get_string('date'))
+            'firstname' => array('field' => 'um.firstname', 'title' => get_string('firstname')),
+            'lastname' => array('field' => 'um.lastname', 'title' => get_string('lastname')),
+            'timemodified' => array('field' => 'q.timemodified', 'title' => get_string('date'))
         );
     }
 }
