+++ question/classes/local/bank/question_name_column.php
@@ -15,63 +15,62 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * A column with a checkbox for each question with name q{questionid}.
+ * A column type for the name of the question name.
  *
  * @package   core_question
  * @copyright 2009 Tim Hunt
+ * @author    2021 Safat Shahin <safatshahin@catalyst-au.net>
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-namespace core_question\bank;
-defined('MOODLE_INTERNAL') || die();
 
-use core\output\checkbox_toggleall;
+namespace core_question\local\bank;
+defined('MOODLE_INTERNAL') || die();
 
 
 /**
- * A column with a checkbox for each question with name q{questionid}.
+ * A column type for the name of the question name.
  *
  * @copyright 2009 Tim Hunt
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class checkbox_column extends column_base {
+class question_name_column extends column_base {
+    protected $checkboxespresent = null;
 
     public function get_name() {
-        return 'checkbox';
+        return 'questionname';
     }
 
     protected function get_title() {
-        global $OUTPUT;
-
-        $mastercheckbox = new checkbox_toggleall('qbank', true, [
-            'id' => 'qbheadercheckbox',
-            'name' => 'qbheadercheckbox',
-            'value' => '1',
-            'label' => get_string('selectall'),
-            'labelclasses' => 'accesshide',
-        ]);
-
-        return $OUTPUT->render($mastercheckbox);
+        return get_string('question');
     }
 
-    protected function get_title_tip() {
-        return get_string('selectquestionsforbulk', 'question');
+    protected function label_for($question) {
+        if (is_null($this->checkboxespresent)) {
+            $this->checkboxespresent = $this->qbank->has_column('core_question\bank\checkbox_column');
+        }
+        if ($this->checkboxespresent) {
+            return 'checkq' . $question->id;
+        } else {
+            return '';
+        }
     }
 
     protected function display_content($question, $rowclasses) {
-        global $OUTPUT;
-
-        $checkbox = new checkbox_toggleall('qbank', false, [
-            'id' => "checkq{$question->id}",
-            'name' => "q{$question->id}",
-            'value' => '1',
-            'label' => get_string('select'),
-            'labelclasses' => 'accesshide',
-        ]);
-
-        echo $OUTPUT->render($checkbox);
+        $labelfor = $this->label_for($question);
+        if ($labelfor) {
+            echo '<label for="' . $labelfor . '">';
+        }
+        echo format_string($question->name);
+        if ($labelfor) {
+            echo '</label>';
+        }
     }
 
     public function get_required_fields() {
-        return array('q.id');
+        return array('q.id', 'q.name');
+    }
+
+    public function is_sortable() {
+        return 'q.name';
     }
 }
