+++ question/classes/local/bank/tags_action_column.php
@@ -19,9 +19,10 @@
  *
  * @package   core_question
  * @copyright 2018 Simey Lameze <simey@moodle.com>
+ * @author    2021 Safat Shahin <safatshahin@catalyst-au.net>
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-namespace core_question\bank;
+namespace core_question\local\bank;
 defined('MOODLE_INTERNAL') || die();
